ToggleClass app.js
Finally got the toggleclass to work on click, so it adds and removes the
front class. There is a scoring issue however, where you have to click
each card twice the first time around. This is because I had to convert
the JS into jQuery and I did it via a additional event listner. Need to
find something else.