App.js rescoped the event listenr
Previous event listener was not working, so removed the respond to click
function and the event listener to the game area, so dropped the
evt.target process.
Added the toggle event listener to each col incvidually. Will figure out
how to improve for performance later on once it works!! :-D